Fully functional cart

+++ Cart/views.py
@@ -2,6 +2,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from .serializers import CartSerializer,CartCreationSerializer
 from .models import Cart, CartItems
+from Core.permissions import CanEditCart
 from rest_framework.generics import CreateAPIView
 from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
@@ -23,7 +24,7 @@ class CartApiView(APIView):
     def post(self,request):
         """This endpoint handles the POST request and requires a list of products to be added to the cart
         """
-        cart = Cart.objects.get_or_create(user=request.user)
+        cart, _ = Cart.objects.get_or_create(user=request.user)
         serializer = CartCreationSerializer(data=request.data, context={'cart' : cart}, many=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -51,9 +52,10 @@ class DeleteFromCartAPIView(APIView):
         The endpoint handles the following request type:
         - DELETE remove a particular item from the cart
     """
-    permission_classes = [IsAuthenticated]
+    permission_classes = [CanEditCart]
     def delete(self, request, pk):
         item = CartItems.objects.get(pk=pk)
+        self.check_object_permissions(request, item)
         item.delete()
         return Response(
             {
@@ -62,9 +64,10 @@ class DeleteFromCartAPIView(APIView):
         )
 
 class ClearCartAPIView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [CanEditCart]
     def delete(self, request):
         items = CartItems.objects.filter(cart__user=request.user)
+        self.check_object_permissions(request, items)
         items.delete()
         return Response(
             {
